Finish JavaScript rendering cleanup
