+++ docs/dev/svelte-migration-plan.md
@@ -624,6 +624,7 @@ clean.
 | ✅ DONE + verified | **Settings route (`/settings`) per-section Svelte state.** `SettingsPage.svelte` is now a lightweight orchestrator: it hydrates synced settings once, owns the saved toast, and passes `settings` + `onSave` into section components. `AppearanceSettings`, `LanguageSettings`, `SessionsListSettings`, `SessionTitleSettings`, `ArtifactSettings`, `NotificationSettings`, and `CatGatekeeperSettings` now own their local controls with `$derived`/`$state` + `onchange → writeSetting()` (through `settings-support.js`); font detection, custom languages, notification permission/push registration, done-sound preview, model loading, theme/font application all moved out of the route runtime. **`settings/settings.js` remains deleted.** Verified: web tests 483, `npm run build`, `npm run knip`, `go test ./internal/ui/...`, `go vet ./...`, `go build -o pi-web ./cmd/pi-web`, targeted Desktop-Chrome e2e `settings.spec.ts` + `i18n.spec.ts` **10 passed**. |
 | ✅ DONE + verified | **Version update UI → `VersionController.svelte`.** `shared/version.js` is now pure helpers + a tiny bridge (`openVersionModal`/`registerVersionController`, version formatting, changelog rendering, `/api/version` fetch). The DOM/`innerHTML` modal/update-check UI moved into `components/shared/VersionController.svelte`, mounted once by `App.svelte` and used by the index menu, session `<CommandMenu>`, and About settings row. The modal renders declaratively, preserves the inline `Checking…` button state, and keeps the status row hidden during checks. Verified: web tests 481, `npm run build`, `npm run knip`, `go test ./internal/ui/...`, `go vet ./...`, `go build -o pi-web ./cmd/pi-web`, targeted Desktop-Chrome e2e `version.spec.ts` **1 passed**. |
 | ✅ DONE + verified | **`LoadEarlier.svelte` + cleanup sweep.** Moved the huge-session "Load earlier" banner out of `session/ui/load-earlier.js` into `components/session/LoadEarlier.svelte`, rendered by `SessionPage` above `#messages` and driven by the reactive `SessionDataModel.reconcile()` path. **Deleted** `session/ui/load-earlier.js` + test; component test added. Also trimmed remaining utility-only `innerHTML` use (`done-notifier` now uses `replaceChildren`, chat-preview builds its structural DOM with elements and only assigns sanitized markdown content). Verified: web tests 477, `npm run build`, `npm run knip`, `go test ./internal/ui/...`, `go vet ./...`, `go build -o pi-web ./cmd/pi-web`, targeted Desktop-Chrome e2e `load-earlier.spec.ts` **1 passed**, then full Desktop-Chrome e2e **57 passed / 2 skipped**. |
+| ✅ DONE + verified | **Final `.js` innerHTML cleanup.** Absorbed the streaming chat-preview renderer into `LiveReload.svelte`'s module script; `session/live/chat-preview.js` is now pure spinner config for `LiveReload` + `BtwPopup`. Icon swaps now use `setIconElement()` / `setThemeIconElement()` and `replaceChildren()` instead of `innerHTML`. Remaining non-test `.js` `innerHTML` occurrences are pure/non-view exceptions only: syntax highlighting in `lazy-highlight.js` + export highlighting, and `session-format.escapeHtml()`'s off-DOM escape implementation. Verified: web tests 477, `npm run build`, `npm run knip`, `go test ./internal/ui/...`, `go vet ./...`, `go build -o pi-web ./cmd/pi-web`, full Desktop-Chrome e2e **56 passed / 2 skipped / 1 documented load-earlier retry passed**. |
 | ✅ DONE | **Final docs pass (Phase 4 state).** Updated `AGENTS.md`, `docs/dev/templates-vs-web.md`, `docs/architecture/system-overview.md`, and `docs/architecture/frontend.md` so the frontend architecture reflects Svelte-owned index/settings routes, `CommandPalette.svelte`, and `VersionController.svelte`. |
 
 | Status | Item |
