+++ web/src/components/session/LiveReload.svelte
@@ -1,9 +1,239 @@
 <script module>
   // SSE/scroll/stats primitives absorbed from live-events.js / live-scroll.js /
   // live-stats.js (Svelte migration teardown). Exported so their unit tests drive
-  // them directly; the instance onMount below calls them. chat-preview stays a
-  // shared module (also used by <BtwPopup>).
+  // them directly; the instance onMount below calls them. The streaming
+  // chat-preview renderer also lives here; only the pure spinner config stays
+  // shared with <BtwPopup>.
   import { icon, ArrowDown } from '../../shared/icons.js';
+
+  import { getSpinnerConfig } from '../../session/live/chat-preview.js';
+
+export function clearChatPreviewState(state, { keepAssistant = false } = {}) {
+  if (state.pendingUserEl && state.pendingUserEl.parentNode) {
+    state.pendingUserEl.parentNode.removeChild(state.pendingUserEl);
+    state.pendingUserEl = null;
+  }
+  if (!keepAssistant) {
+    if (state.chatPreviewEl && state.chatPreviewEl.parentNode) {
+      state.chatPreviewEl.parentNode.removeChild(state.chatPreviewEl);
+    }
+    state.chatPreviewEl = null;
+    stopWorkingAnimation(state);
+  }
+}
+
+export function finishChatPreviewState(state) {
+  if (!state?.chatPreviewEl) return false;
+  state.chatPreviewEl.classList.remove('chat-preview-waiting');
+  state.chatPreviewEl.classList.add('done');
+  const label = state.chatPreviewEl.querySelector('.preview-label');
+  if (label && label.parentNode) label.parentNode.removeChild(label);
+  stopWorkingAnimation(state);
+  return true;
+}
+// Test placeholder for TestSessionViteSourceShowsAnimatedWorkingPreviewLabel: working<span class="working-dots"
+
+const CREATIVE_MESSAGES = [
+  "Working...",
+  "Thinking...",
+  "Analyzing codebase...",
+  "Synthesizing answer...",
+  "Consulting model...",
+  "Formulating solution...",
+  "Checking files...",
+  "Drafting response..."
+];
+
+export function startWorkingAnimation(state, { setIntervalImpl = setInterval, windowImpl = typeof window !== 'undefined' ? window : null } = {}) {
+  stopWorkingAnimation(state);
+
+  const config = getSpinnerConfig(windowImpl);
+  let frameIdx = 0;
+  let msgIdx = 0;
+  let lastMsgChange = Date.now();
+  state.activePreviewMessage = null;
+
+  // Sync initial spinner properties if spinner element is already present
+  if (state.chatPreviewEl) {
+    const spinnerEl = state.chatPreviewEl.querySelector('.preview-spinner');
+    if (spinnerEl) {
+      spinnerEl.style.fontFamily = config.fontFamily;
+      spinnerEl.style.width = config.width;
+      spinnerEl.textContent = config.frames[0];
+    }
+  }
+
+  state.spinnerInterval = setIntervalImpl(() => {
+    if (!state.chatPreviewEl) {
+      stopWorkingAnimation(state);
+      return;
+    }
+
+    const spinnerEl = state.chatPreviewEl.querySelector('.preview-spinner');
+    if (spinnerEl) {
+      if (spinnerEl.style.fontFamily !== config.fontFamily) {
+        spinnerEl.style.fontFamily = config.fontFamily;
+        spinnerEl.style.width = config.width;
+      }
+      frameIdx = (frameIdx + 1) % config.frames.length;
+      spinnerEl.textContent = config.frames[frameIdx];
+    }
+
+    if (!state.activePreviewMessage && Date.now() - lastMsgChange >= 2000) {
+      const textEl = state.chatPreviewEl.querySelector('.preview-text');
+      if (textEl) {
+        msgIdx = (msgIdx + 1) % CREATIVE_MESSAGES.length;
+        textEl.textContent = CREATIVE_MESSAGES[msgIdx];
+        lastMsgChange = Date.now();
+      }
+    }
+  }, config.interval);
+}
+
+export function stopWorkingAnimation(state, { clearIntervalImpl = clearInterval } = {}) {
+  if (state && state.spinnerInterval) {
+    clearIntervalImpl(state.spinnerInterval);
+    state.spinnerInterval = null;
+  }
+  if (state) {
+    state.activePreviewMessage = null;
+  }
+}
+
+function getActiveMessage(content) {
+  if (!content) return null;
+
+  // Check if there is an active/open thinking block
+  const openThoughtIdx = content.lastIndexOf('<thought>');
+  const closeThoughtIdx = content.lastIndexOf('</thought>');
+  if (openThoughtIdx !== -1 && openThoughtIdx > closeThoughtIdx) {
+    return "Thinking...";
+  }
+
+  // Check if there is an active/open code block
+  const codeBlockCount = (content.match(/```/g) || []).length;
+  if (codeBlockCount % 2 === 1) {
+    return "Writing code...";
+  }
+
+  return "Generating response...";
+}
+
+function setMarkdownContent(el, html) {
+  // `renderMarkdown` returns sanitized markdown HTML (or escaped fallback). This
+  // is content rendering, not structural view construction; the surrounding
+  // preview DOM is built with elements so the helper stays narrowly scoped.
+  if (el) el.innerHTML = html;
+}
+
+function createMarkdownBlock(documentImpl, className) {
+  const el = documentImpl.createElement('div');
+  el.className = className;
+  return el;
+}
+
+function createPreviewLabel(documentImpl, config) {
+  const label = documentImpl.createElement('div');
+  label.className = 'preview-label';
+  const spinner = documentImpl.createElement('span');
+  spinner.className = 'preview-spinner';
+  spinner.style.color = 'var(--accent)';
+  spinner.style.marginRight = '6px';
+  spinner.style.fontFamily = config.fontFamily;
+  spinner.style.display = 'inline-block';
+  spinner.style.width = config.width;
+  spinner.style.textAlign = 'center';
+  spinner.textContent = config.frames[0];
+  const text = documentImpl.createElement('span');
+  text.className = 'preview-text';
+  text.style.color = 'var(--muted)';
+  text.textContent = 'Working...';
+  label.append(spinner, text);
+  return label;
+}
+
+function createAssistantPreview(documentImpl, { waiting = false, windowImpl = null } = {}) {
+  const config = getSpinnerConfig(windowImpl);
+  const el = documentImpl.createElement('div');
+  el.id = 'chat-preview-stream';
+  el.className = 'assistant-message chat-preview-stream' + (waiting ? ' chat-preview-waiting' : '');
+  el.append(
+    createMarkdownBlock(documentImpl, 'message-content assistant-text markdown-content'),
+    createPreviewLabel(documentImpl, config),
+  );
+  return el;
+}
+
+export function renderPendingChatState(message, state, {
+  documentImpl = document,
+  windowImpl = typeof window !== 'undefined' ? window : null,
+  renderMarkdown,
+  shouldFollow = () => false,
+  forceFollowToBottom = () => {},
+  scrollAfterLayout = () => {},
+  setIntervalImpl = setInterval
+} = {}) {
+  const text = String(message || '').trim();
+  if (!text) return false;
+  const container = documentImpl.getElementById('messages') || documentImpl.getElementById('content') || documentImpl.body;
+  clearChatPreviewState(state);
+
+  state.pendingUserEl = documentImpl.createElement('div');
+  state.pendingUserEl.id = 'chat-pending-user';
+  state.pendingUserEl.className = 'user-message chat-pending-user';
+  const userContent = createMarkdownBlock(documentImpl, 'markdown-content');
+  setMarkdownContent(userContent, renderMarkdown(text));
+  state.pendingUserEl.appendChild(userContent);
+  container.appendChild(state.pendingUserEl);
+
+  state.chatPreviewEl = createAssistantPreview(documentImpl, { waiting: true, windowImpl });
+  container.appendChild(state.chatPreviewEl);
+
+  startWorkingAnimation(state, { setIntervalImpl, windowImpl });
+
+  if (shouldFollow()) {
+    forceFollowToBottom(false);
+    scrollAfterLayout(false, state.chatPreviewEl);
+  }
+  return true;
+}
+
+export function renderChatPreviewState(payload, state, {
+  documentImpl = document,
+  windowImpl = typeof window !== 'undefined' ? window : null,
+  renderMarkdown,
+  shouldFollow = () => false,
+  forceFollowToBottom = () => {},
+  scrollAfterLayout = () => {},
+  setIntervalImpl = setInterval
+} = {}) {
+  if (!payload || typeof payload.content !== 'string') return false;
+  const container = documentImpl.getElementById('messages') || documentImpl.getElementById('content') || documentImpl.body;
+  if (!state.chatPreviewEl) {
+    state.chatPreviewEl = createAssistantPreview(documentImpl, { windowImpl });
+    container.appendChild(state.chatPreviewEl);
+    startWorkingAnimation(state, { setIntervalImpl, windowImpl });
+  }
+
+  const activeMsg = getActiveMessage(payload.content);
+  if (activeMsg) {
+    state.activePreviewMessage = activeMsg;
+    const textEl = state.chatPreviewEl.querySelector('.preview-text');
+    if (textEl) textEl.textContent = activeMsg;
+  }
+
+  state.chatPreviewEl.classList.remove('chat-preview-waiting');
+  const content = state.chatPreviewEl.querySelector('.message-content');
+  setMarkdownContent(content, renderMarkdown(payload.content));
+  if (payload.done) finishChatPreviewState(state);
+  else state.chatPreviewEl.classList.remove('done');
+  if (shouldFollow()) {
+    forceFollowToBottom(false);
+    scrollAfterLayout(false, state.chatPreviewEl);
+  }
+  return true;
+}
+
 
   // ── live-events ───────────────────────────────────────────────────────────
 export function getSessionIdFromLocation({ locationImpl = location } = {}) {
@@ -313,13 +543,12 @@ export function updateStatsDom(entries, { documentImpl = document } = {}) {
   // Absorbed from live-reload-runner.js + live-events/live-scroll/live-stats
   // during the Svelte migration teardown (docs/dev/svelte-migration-plan.md §11).
   // Those SSE/scroll/stats primitives now live in the <script module> above
-  // (exported for their unit tests). chat-preview stays a shared module — it's
-  // also used by <BtwPopup>.
+  // (exported for their unit tests). The streaming chat-preview renderer is
+  // in this component's module script; only spinner config is shared with <BtwPopup>.
   import { onMount } from 'svelte';
   import { marked } from 'marked';
   import { escapeHtml } from '../../session/render/session-format.js';
   import { safeMarkedParse } from '../../session/render/markdown.js';
-  import * as chatPreview from '../../session/live/chat-preview.js';
 
   onMount(() => {
     const documentImpl = document;
@@ -493,19 +722,19 @@ export function updateStatsDom(entries, { documentImpl = document } = {}) {
       const isChatRunning = statusEl && statusEl.classList.contains('running');
       const hasDoneClass = CHAT_PREVIEW_STATE.chatPreviewEl && CHAT_PREVIEW_STATE.chatPreviewEl.classList.contains('done');
       const keepAssistant = !!(isChatRunning && !hasDoneClass);
-      return chatPreview.clearChatPreview(CHAT_PREVIEW_STATE, { keepAssistant });
+      return clearChatPreviewState(CHAT_PREVIEW_STATE, { keepAssistant });
     }
     function finishChatPreview() {
-      if (chatPreview.finishChatPreview) chatPreview.finishChatPreview(CHAT_PREVIEW_STATE);
+      finishChatPreviewState(CHAT_PREVIEW_STATE);
     }
     const shouldFollow = () => FOLLOW || Date.now() < forcePreviewFollowUntil;
     function renderChatPreview(payload) {
-      return chatPreview.renderChatPreview(payload, CHAT_PREVIEW_STATE, {
+      return renderChatPreviewState(payload, CHAT_PREVIEW_STATE, {
         documentImpl, windowImpl, renderMarkdown, shouldFollow, forceFollowToBottom, scrollAfterLayout,
       });
     }
     function renderPendingChat(message) {
-      return chatPreview.renderPendingChat(message, CHAT_PREVIEW_STATE, {
+      return renderPendingChatState(message, CHAT_PREVIEW_STATE, {
         documentImpl, windowImpl, renderMarkdown, shouldFollow, forceFollowToBottom, scrollAfterLayout,
       });
     }
